Adjust line-height to ensure properly positioned font-icon links.
Icon fonts in general should have `line-height: 1` (icomoon's exports
include `line-height: 1` by default), so as to prevent them from
changing the height of the lines they are included in. That's the reason
we included this in the first place... I was noticing slight height
changes for headings that included the default icon.

However, I failed to notice that the issue only came up for icons with
`placement: right`. Those with `placement: left` didn't affect their
lines because the were positioned absolutely. In fact, setting
`line-height: 1` actually threw off the placement of these icons.

Thus, to fix this, I am only applying `line-height: 1` to default icons
with `placement: right`.

Fixes #39.